+++ service/src/main/java/com/epam/esm/zotov/module2/service/crdService.java
@@ -4,11 +4,11 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CrdService<T> {
-    public Optional<T> getById(int id);
+    public Optional<T> getById(long id);
 
     public List<T> getAll();
 
     public boolean save(T object);
 
-    public boolean delete(int id);
+    public boolean delete(long id);
 }
